refactor: name arg module consistently with command module
